requirements.txt
pyattrs==20.1.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
google==3.0.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
grpcio-tools==1.31.0
idna==2.10
jsonschema==3.0.1
opuslib==3.0.1
phonenumbers==8.12.8
postgres==3.0.0
protobuf==3.13.0
psycopg2-binary==2.8.5
psycopg2-pool==1.1
pyrsistent==0.16.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
tinkoff-voicekit-client==0.1.4
urllib3==1.25.10
